Added missing folder fallbacks
The script will now check if the destination folder in copy and move operations exists. If the folder does not exist, the script will create it before continuing its runtime after informing the user. This will prevent errors with the script creating a merged file with no extension when the folder specified does not exist.